refactor(asn1-parser): wip
