+++ crates/asn1-parser/src/lib.rs
@@ -8,7 +8,7 @@ mod macros;
 mod asn1;
 mod constructors;
 mod error;
-mod generic_types;
+mod primitives;
 mod length;
 mod reader;
 mod string;
@@ -20,7 +20,7 @@ mod writer;
 pub use asn1::{Asn1, Asn1Type, OwnedAsn1, OwnedAsn1Type, OwnedRawAsn1EntityData, RawAsn1EntityData};
 pub use constructors::*;
 pub use error::Error;
-pub use generic_types::*;
+pub use primitives::*;
 use reader::Reader;
 pub use string::*;
 pub use tag::Tag;
